Aggiustata la delete. Aggiunta conferma tasto "elimina". Eliminato footer in "editPrezzi". Effettuato "reformat code" su tutto il progetto. Aggiustare la print nell'index.

+++ Functions/printOfferta.php
@@ -14,27 +14,28 @@ while ($rowC = mysqli_fetch_array($result)) {
     //Formattano rispettivamente la data inizio e la data fine così gg/mm/aaaa
     $temp1 = explode(" ", $rowC['dataInizio']);
     $temp2 = explode("-", $temp1[0]);
-    $dataInizioFormattata = $temp2[2]."/".$temp2[1]."/".$temp2[0];
+    $dataInizioFormattata = $temp2[2] . "/" . $temp2[1] . "/" . $temp2[0];
 
     $temp1 = explode(" ", $rowC['dataFine']);
     $temp2 = explode("-", $temp1[0]);
-    $dataFineFormattata = $temp2[2]."/".$temp2[1]."/".$temp2[0];
+    $dataFineFormattata = $temp2[2] . "/" . $temp2[1] . "/" . $temp2[0];
 
 
     echo '<td>' . $rowC['nome'] . '</td>';
-    echo '<td>' ; echo $dataInizioFormattata;  '</td>';
-    echo '<td>' ; echo $dataFineFormattata; '</td>';
+    echo '<td>';
+    echo $dataInizioFormattata;
+    '</td>';
+    echo '<td>';
+    echo $dataFineFormattata;
+    '</td>';
     echo '<td>' . $rowC['descrizione'] . '</td>';
     echo '<td>' . $rowC['prezzo'] . '</td>';
 
-    echo '<td> <a href="deleteOfferta.php?id='.$rowC['id'].'"> Elimina </a> </td>';
+    echo '<td> <a onclick="return confirm(\'Confermi?\');" href="/spa/Functions/deleteOfferta.php?id=' . $rowC['id'] . '"> Elimina </a> </td>';
 
     echo "</tr>";
 }
 echo "</table>";
-
-
-
 
 
 ?>
